pyproject.toml
[project]
name = "insta_scraping"
version = "0.1.0"
description = ""
authors = [
    {name = "Douglas Oliveira",email = "63203428+imdsoliveira@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "taskipy (>=1.14.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "instaloader (>=4.14.1,<5.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "python get_profile.py"
update = "python.exe -m pip install --upgrade pip && poetry update"